move initialise and terminate block device driver methods to device_plugin action_plugin to be loaded when agent starts in monitored mode
