+++ chroma-agent/chroma_agent/action_plugins/manage_node.py
@@ -9,13 +9,6 @@ from chroma_agent.lib.shell import AgentShell
 from chroma_agent.log import console_log
 from chroma_agent.device_plugins.action_runner import CallbackAfterResponse
 from chroma_agent.lib.pacemaker import PacemakerConfig
-from iml_common.blockdevices.blockdevice import BlockDevice
-from iml_common.lib import util
-from iml_common.lib.agent_rpc import agent_error
-from iml_common.lib.agent_rpc import agent_result_ok
-from chroma_agent.lib.agent_startup_functions import agent_daemon_startup_function
-from chroma_agent.lib.agent_teardown_functions import agent_daemon_teardown_function
-from chroma_agent import config
 
 
 def ssi(runlevel):
@@ -62,32 +55,4 @@ def reboot_server(at_time = "now"):
 
     raise CallbackAfterResponse(None, _reboot)
 
-
-# When the agent is run we want to allow block devices to do any initialization that they might need, this function
-# may also be called by the manager.
-@agent_daemon_startup_function()
-def initialise_block_device_drivers():
-    console_log.info("Initialising drivers for block device types")
-    for cls in util.all_subclasses(BlockDevice):
-        error = cls.initialise_driver(config.profile_managed)
-
-        if error:
-            return agent_error(error)
-
-    return agent_result_ok
-
-
-# When the agent is stopped we want to allow block devices to do any termination that they might need, this function
-# may also be called by the manager.
-@agent_daemon_teardown_function()
-def terminate_block_device_drivers():
-    console_log.info("Terminating drivers for block device types")
-    for cls in util.all_subclasses(BlockDevice):
-        error = cls.terminate_driver()
-
-        if error:
-            return agent_error(error)
-
-    return agent_result_ok
-
-ACTIONS = [reboot_server, shutdown_server, fail_node, stonith, initialise_block_device_drivers]
+ACTIONS = [reboot_server, shutdown_server, fail_node, stonith]
